pyproject.toml
[tool.poetry]
name = "tataro-project"
version = "0.1.0"
description = ""
authors = ["wkdsh21 <wkdsh21@naver.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
django-environ = "^0.12.0"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"
drf-yasg-stubs = "^0.1.4"
channels-redis = "^4.2.1"
django-stubs = "^5.1.2"
djangorestframework-stubs = "^3.15.2"
psycopg2 = "^2.9.10"
boto3 = "1.35.76"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
django-cors-headers = "^4.7.0"
djangorestframework-simplejwt = "^5.4.0"
python-dotenv = "^1.0.1"
daphne = "3.0.2"
channels = "3.0.5"
portone-server-sdk = "^0.13.0"



[tool.poetry.group.dev.dependencies]
mypy = "^1.14.1"
isort = "^5.13.2"
black = "^24.10.0"
django-stubs = {extras = ["compatible-mypy"], version = "^5.1.2"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
python_version = 3.12
strict = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.base"